requirements.txt
fastapi
uvicorn
aiohttp
pycryptodome